requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pillow==12.1.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
smmap==5.0.2
